Add Clash.Normalize.Transformations.ANF
All transformations concerning conversion to ANF now live in a
separate module. This is part of work to make the normalization
modules quicker to compile.
